backend/requirements.txt
flask==2.3.2
python-dotenv==1.0.0
google-generativeai==0.3.2
numpy==1.24.3
requests==2.31.0